move utils functions to utils file
